LayoutTests/platform/ios/imported/w3c/web-platform-tests/mathml/relations/css-styling/width-height-006-expected.txt


PASS mrow
PASS mrow (vertical overflow)
FAIL mpadded assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL mpadded (vertical overflow) assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL msqrt assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL msqrt (vertical overflow) assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL mroot assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL mroot (vertical overflow) assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL munder assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL munder (vertical overflow) assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL mover assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL mover (vertical overflow) assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL munderover assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL munderover (vertical overflow) assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL msub assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL msub (vertical overflow) assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL msup assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL msup (vertical overflow) assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL msubsup assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL msubsup (vertical overflow) assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL mmultiscripts assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL mmultiscripts (vertical overflow) assert_greater_than: height is modified expected a number greater than 1 but got 0
PASS mtext
PASS mtext (vertical overflow)
FAIL legacy menclose assert_greater_than: height is modified expected a number greater than 1 but got 0
FAIL legacy menclose (vertical overflow) assert_greater_than: height is modified expected a number greater than 1 but got 0
X
X
X

X
X
X

X
X
X
X
X
X

X
X
X

X
X
X
X
X
X

X
X
X

X
X
X
X
X

X
X

X
X
X
X

X
X

X
X
X
X

X
X

X
X
X
X
X

X
X
X

X
X
X
X
X

X
X

X
X
X
X

X
X

X
X
X
X
X

X
X
X

X
X
X
X
X
X
X
X

X
X
X
X
X

X
X
X
X
X
X X X

X X X

X X X
X
X
X

X
X
X

X
X
X
